aggregated heuristic controller tests into one file
